Move cards tags logic to discrete attributes
The objects and locations are now attributes on the card
and will be used for matching previously stored cards in
the Card deck instead of tags.

As a temporary measure, recognized object: & location:
tags on former cards will be added as attributes
appropriately.

Also includes various fixes and changes in the UI.
Add missing Javadoc to pass the checkstyle checks.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>