Add unobstrusive "Install App" button
Add the "Install App" button to promote A2HS when the PWA
can be added as an app to the homescreen or desktop.

Available on Chrome 70+ for Windows, Android or Chrome OS.
Available through chrome:flags (PWA, PWA link capturing)
on macOS and Linux - expected to land on Chrome 72.

References:
https://developers.google.com/web/progressive-web-apps/desktop
https://developers.google.com/web/fundamentals/app-install-banners/

Signed-off-by: Yannick Schaus <habpanel@schaus.net>